Create LICENSE
This license permits anyone to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the software. It also requires that the original copyright notice and permission notice be included in all copies or substantial portions of the software. 

The license ensures the software is provided "as is," without any warranties, offering clarity to both contributors and users about their rights and responsibilities.